app修改若干 server添加调拨单的小程序接口

+++ WS-KingdeeOnlyCode/app/src/main/java/com/fangzuo/assist/Activity/OutsourcingOrdersOSActivity.java
@@ -22,6 +22,7 @@ import android.widget.Spinner;
 import android.widget.TextView;
 
 import com.fangzuo.assist.ABase.BaseActivity;
+import com.fangzuo.assist.Activity.Crash.App;
 import com.fangzuo.assist.Adapter.BatchNoSpAdapter;
 import com.fangzuo.assist.Adapter.CheckBatchAdapter;
 import com.fangzuo.assist.Adapter.EmployeeSpAdapter;
@@ -52,6 +53,7 @@ import com.fangzuo.assist.Dao.Unit;
 import com.fangzuo.assist.Dao.Wanglaikemu;
 import com.fangzuo.assist.Dao.WaveHouse;
 import com.fangzuo.assist.R;
+import com.fangzuo.assist.RxSerivce.MySubscribe;
 import com.fangzuo.assist.Utils.Asynchttp;
 import com.fangzuo.assist.Utils.BasicShareUtil;
 import com.fangzuo.assist.Utils.CommonMethod;
@@ -1222,13 +1224,15 @@ public class OutsourcingOrdersOSActivity extends BaseActivity {
         PurchaseInStoreUploadBean pBean = new PurchaseInStoreUploadBean();
         pBean.list = data;
         Gson gson = new Gson();
-        Asynchttp.post(mContext, getBaseUrl() + WebApi.PushDownOCOSUpload, gson.toJson(pBean), new Asynchttp.Response() {
+        App.getRService().doIOAction(WebApi.PushDownOCOSUpload, gson.toJson(pBean), new MySubscribe<CommonResponse>() {
             @Override
-            public void onSucceed(CommonResponse cBean, AsyncHttpClient client) {
+            public void onNext(CommonResponse commonResponse) {
+                super.onNext(commonResponse);
+                if (!commonResponse.state)return;
                 MediaPlayer.getInstance(mContext).ok();
                 Toast.showText(mContext, "上传成功");
-                    t_detailDao.deleteInTx(t_detailDao.queryBuilder().where(T_DetailDao.Properties.Activity.eq(activity)).build().list());
-                    t_mainDao.deleteInTx(t_mainDao.queryBuilder().where(T_mainDao.Properties.Activity.eq(activity)).build().list());
+                t_detailDao.deleteInTx(t_detailDao.queryBuilder().where(T_DetailDao.Properties.Activity.eq(activity)).build().list());
+                t_mainDao.deleteInTx(t_mainDao.queryBuilder().where(T_mainDao.Properties.Activity.eq(activity)).build().list());
                 for (int i = 0; i < fidc.size(); i++) {
                     List<PushDownSub> pushDownSubs = pushDownSubDao.queryBuilder().where(
                             PushDownSubDao.Properties.Tag.eq(tag),
@@ -1251,8 +1255,9 @@ public class OutsourcingOrdersOSActivity extends BaseActivity {
             }
 
             @Override
-            public void onFailed(String Msg, AsyncHttpClient client) {
-                Toast.showText(mContext, Msg);
+            public void onError(Throwable e) {
+                super.onError(e);
+                Toast.showText(mContext, e.getMessage());
                 btnBackorder.setClickable(true);
                 MediaPlayer.getInstance(mContext).error();
                 LoadingUtil.dismiss();
